Oscillators tracking partials
Took a stab at making the oscillators track the partials. The way it
works is that each peak has a new category for oscillator (along with
partial), and as we go from peak to peak in a partial, we pass the
oscillator along (sort of like a linked list thing). Not convinced that
this is the best way to implement this, but it kinda works.